requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy>=1.21.0
RPi.GPIO>=0.7.1
